add a flag to allow excluding products from points
In order to do this I've had to add another field to spree_products
that marks a product as needing to track loyalty points.

I've had to switch from using order.item_total to a new construct of
loyalty_point_eligible_total which touches a lot of the extension but is
unavoidable.

however, this work will be required by the next step of excluding
promotions from point calculations.